+++ src/main/java/org/torproject/onionoo/server/ResourceServlet.java
@@ -68,7 +68,7 @@ public class ResourceServlet extends HttpServlet {
   private static Set<String> knownParameters = new HashSet<>(
       Arrays.asList(("type,running,search,lookup,fingerprint,country,as,"
           + "flag,first_seen_days,last_seen_days,contact,order,limit,"
-          + "offset,fields,family,version").split(",")));
+          + "offset,fields,family,version,host_name").split(",")));
 
   private static Set<String> illegalSearchQualifiers =
       new HashSet<>(Arrays.asList(("search,fingerprint,order,limit,"
@@ -291,6 +291,15 @@ public class ResourceServlet extends HttpServlet {
         return;
       }
       rh.setVersion(versionParameter);
+    }
+    if (parameterMap.containsKey("host_name")) {
+      String hostNameParameter = this.parseHostNameParameter(
+          parameterMap.get("host_name"));
+      if (null == hostNameParameter) {
+        response.sendError(HttpServletResponse.SC_BAD_REQUEST);
+        return;
+      }
+      rh.setHostName(hostNameParameter);
     }
     if (parameterMap.containsKey("order")) {
       String[] order = this.parseOrderParameter(parameterMap.get("order"));
@@ -580,6 +589,17 @@ public class ResourceServlet extends HttpServlet {
   private String parseVersionParameter(String parameter) {
     if (!versionParameterPattern.matcher(parameter).matches()) {
       /* Version contains illegal character(s). */
+      return null;
+    }
+    return parameter;
+  }
+
+  private static Pattern hostNameParameterPattern =
+      Pattern.compile("^[0-9A-Za-z_\\.\\-]+$");
+
+  private String parseHostNameParameter(String parameter) {
+    if (!hostNameParameterPattern.matcher(parameter).matches()) {
+      /* Host name contains illegal character(s). */
       return null;
     }
     return parameter;