+++ src/test/java/org/torproject/onionoo/server/ResourceServletTest.java
@@ -143,7 +143,7 @@ public class ResourceServletTest {
             "0025C136C1F3A9EEFE2AE3F918F03BFA21B5070B" })),
         new TreeSet<>(Arrays.asList(
         new String[] { "001C13B3A55A71B977CA65EC85539D79C653A3FC" })),
-        "0.2.3.25");
+        "0.2.3.25", "ppp-62-216-201-221.dynamic.mnet-online.de");
     org.torproject.onionoo.docs.SummaryDocument relayFerrari458 =
         new org.torproject.onionoo.docs.SummaryDocument(true, "Ferrari458",
         "001C13B3A55A71B977CA65EC85539D79C653A3FC", Arrays.asList(
@@ -155,7 +155,8 @@ public class ResourceServletTest {
         new TreeSet<String>(Arrays.asList(new String[] {
             "000C5F55BD4814B917CC474BD537F1A3B33CCE2A" })),
         new TreeSet<>(Arrays.asList(new String[] {
-            "000C5F55BD4814B917CC474BD537F1A3B33CCE2A" })), null);
+            "000C5F55BD4814B917CC474BD537F1A3B33CCE2A" })), null,
+            "c-68-38-171-200.hsd1.in.comcast.net");
     this.relays = new TreeMap<>();
     this.relays.put("000C5F55BD4814B917CC474BD537F1A3B33CCE2A",
         relayTorkaZ);
@@ -171,7 +172,7 @@ public class ResourceServletTest {
         DateTimeHelper.parse("2013-04-16 18:00:00"), "AS6830",
         "1024d/51e2a1c7 \"steven j. murdoch\" "
         + "<tor+steven.murdoch@cl.cam.ac.uk> <fb-token:5sr_k_zs2wm=>",
-        new TreeSet<String>(), new TreeSet<String>(), "0.2.3.25");
+        new TreeSet<String>(), new TreeSet<String>(), "0.2.3.25", null);
     this.relays.put("0025C136C1F3A9EEFE2AE3F918F03BFA21B5070B",
         relayTimMayTribute);
     this.bridges = new TreeMap<>();
@@ -182,7 +183,7 @@ public class ResourceServletTest {
         DateTimeHelper.parse("2013-04-21 18:07:03"), false,
         new TreeSet<>(Arrays.asList(new String[] { "Valid" })), -1L,
         null, DateTimeHelper.parse("2013-04-20 15:37:04"), null, null,
-        null, null, null);
+        null, null, null, null);
     this.bridges.put("0000831B236DFF73D409AD17B40E2A728A53994F",
         bridgeec2bridgercc7f31fe);
     org.torproject.onionoo.docs.SummaryDocument bridgeUnnamed =
@@ -192,7 +193,7 @@ public class ResourceServletTest {
         DateTimeHelper.parse("2013-04-20 17:37:04"), false,
         new TreeSet<>(Arrays.asList(new String[] { "Valid" })), -1L,
         null, DateTimeHelper.parse("2013-04-14 07:07:05"), null, null,
-        null, null, null);
+        null, null, null, null);
     this.bridges.put("0002D9BDBBC230BD9C78FF502A16E0033EF87E0C",
         bridgeUnnamed);
     org.torproject.onionoo.docs.SummaryDocument bridgegummy =
@@ -203,7 +204,7 @@ public class ResourceServletTest {
         new TreeSet<>(Arrays.asList(new String[] { "Running",
             "Valid" })), -1L, null,
         DateTimeHelper.parse("2013-01-16 21:07:04"), null, null, null,
-        null, null);
+        null, null, null);
     this.bridges.put("1FEDE50ED8DBA1DD9F9165F78C8131E4A44AB756",
         bridgegummy);
   }
@@ -1622,6 +1623,72 @@ public class ResourceServletTest {
   public void testVersionStart() {
     /* This is also correct when comparing strings. */
     this.assertErrorStatusCode("/summary?version=*", 400);
+  }
+
+  @Test
+  public void testHostNameDe() {
+    this.assertSummaryDocument("/summary?host_name=de", 1, null, 0, null);
+  }
+
+  @Test
+  public void testHostNameE() {
+    this.assertSummaryDocument("/summary?host_name=e", 1, null, 0, null);
+  }
+
+  @Test
+  public void testHostNameDotDe() {
+    this.assertSummaryDocument("/summary?host_name=.de", 1, null, 0, null);
+  }
+
+  @Test
+  public void testHostNameOnlineDe() {
+    this.assertSummaryDocument("/summary?host_name=online.de", 1, null, 0,
+        null);
+  }
+
+  @Test
+  public void testHostNameOnlineDeSomeCapitalized() {
+    this.assertSummaryDocument("/summary?host_name=onLiNe.dE", 1, null, 0,
+        null);
+  }
+
+  @Test
+  public void testHostNameOnline() {
+    this.assertSummaryDocument("/summary?host_name=online", 0, null, 0, null);
+  }
+
+  @Test
+  public void testHostNameTorkaZFull() {
+    this.assertSummaryDocument(
+        "/summary?host_name=ppp-62-216-201-221.dynamic.mnet-online.de",
+        1, null, 0, null);
+  }
+
+  @Test
+  public void testHostNameTorkaZSub() {
+    this.assertSummaryDocument(
+        "/summary?host_name=sub.ppp-62-216-201-221.dynamic.mnet-online.de",
+        0, null, 0, null);
+  }
+
+  @Test
+  public void testHostNameNet() {
+    this.assertSummaryDocument("/summary?host_name=net", 1, null, 0, null);
+  }
+
+  @Test
+  public void testHostNameCom() {
+    this.assertSummaryDocument("/summary?host_name=com", 0, null, 0, null);
+  }
+
+  @Test
+  public void testHostNameDollar() {
+    this.assertErrorStatusCode("/summary?host_name=$", 400);
+  }
+
+  @Test
+  public void testHostNameUmlaut() {
+    this.assertErrorStatusCode("/summary?host_name=äöü", 400);
   }
 }